update
added a lot of stuff tonight and it took me a while, i finished off all of the departments and expanded on the call message so it has a random call id, a description, the location of course and i added a blip which automatically sets a waypoint to the call and disappears after 30 seconds. I will change this eventually to you will be able to click Y or N to accept or decline the 000 call. I also added a phone animation and department checking.